Extracted we_opt.h from edit.h
