+++ we_main.h
@@ -7,6 +7,7 @@
 #include "we_fl_unix.h"
 #include "we_fl_fkt.h"
 #include "we_hfkt.h"
+#include "we_opt.h"
 
 void e_ini_desk(ECNT * cn);
 void FARBE_Init(FARBE * fb);
