+++ we_menue.h
@@ -4,6 +4,7 @@
 #include "we_block.h"
 #include "we_edit.h"
 #include "we_mouse.h"
+#include "we_opt.h"
 
 /*   we_menue.c   */
 int WpeHandleMainmenu(int n, FENSTER *f);
