+++ we_prog.h
@@ -2,6 +2,7 @@
 #define WE_MAIN_H
 
 #include "we_edit.h"
+#include "we_hfkt.h"
 #include "we_mouse.h"
 
 #endif
